% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_package0.1.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create a state map with FARS accident locations in a specific year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Integer number indicating the state you want to plot}

\item{year}{the year of which you want to view the FARS data. The year will be coerced to an integer}
}
\value{
This function creates a map with the FARS-locations plotted. Returns a nULL
}
\description{
This function creates a map for the state that is selected. On the map all the locations of accidents of the selected year are plotted.
}
\details{
Make sure that the \code{state.num} argument refers to a state that exists.
}
\examples{
\dontrun{
fars_map_state(state.num=5,year = 2015)
}
}

