% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_package0.1.R
\name{fars_read}
\alias{fars_read}
\title{Read a datafile from the Fatality Analysis Reporting System}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string with the filename. The file should be in csv-format. Zipped csv-files can also be read}
}
\value{
Returns a tibble with all the data from the csv file
}
\description{
This function reads a csv datafile. Though designed speficially to read FARS data it can be used to read all kinds of csv files. It uses the \code{readr::read_csv}) function
}
\details{
If the file doesn't exists it will throw an error.
}
\examples{
\dontrun{
fars_read("accident_2015.csv.bz2"))
fars_read(make_filename(year=2015))
}

}

