% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_package0.1.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Function to create a summary of FARS data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{, a vector with the years of FARS you want to summarize. The years will be coerced to integers}
}
\value{
returns a tibble with the number of FARS observation per month for the provided years
}
\description{
This function creates a summary of FARS datafor the selected years. The summary consists of the count the monthly observation per year.
}
\examples{
\dontrun{
#' fars_summarize_years(c(2013,2014,2015))
}
}

